Use temporary skip_traffic_test in duthost_utils
